hw7-done

+++ src/services/contacts.js
@@ -31,8 +31,6 @@ export const getContacts = async ({
     .merge(contactQuery.getQuery())
     .countDocuments();
 
-  console.log(totalContacts);
-
   const paginationData = calculatePaginationData({
     page,
     perPage,
@@ -47,12 +45,17 @@ export const getContacts = async ({
     ...paginationData,
   };
 };
-export const getContact = (id) => ContactCollection.findById(id);
+export const getContact = (filter) =>
+  ContactCollection.findOne({ _id: filter.contactId, userId: filter.userId });
 export const addContact = (payload) => ContactCollection.create(payload);
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (
+  { contactId, userId },
+  payload,
+  options = {},
+) => {
   const rawResult = await ContactCollection.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, userId },
     payload,
     {
       new: true,
@@ -68,5 +71,5 @@ export const updateContact = async (contactId, payload, options = {}) => {
   };
 };
 
-export const deleteContact = (contactId) =>
-  ContactCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = ({ contactId, userId }) =>
+  ContactCollection.findOneAndDelete({ _id: contactId, userId });
